refactor: Order, OrderTable 간에 단방향 간접참조관계로 변경
